move the command init to a separate file
